+++ elements/src/components.d.ts
@@ -99,6 +99,10 @@ export interface PosAddLiteralValueCustomEvent<T> extends CustomEvent<T> {
     detail: T;
     target: HTMLPosAddLiteralValueElement;
 }
+export interface PosAppCustomEvent<T> extends CustomEvent<T> {
+    detail: T;
+    target: HTMLPosAppElement;
+}
 export interface PosAppDocumentViewerCustomEvent<T> extends CustomEvent<T> {
     detail: T;
     target: HTMLPosAppDocumentViewerElement;
@@ -218,7 +222,20 @@ declare global {
         prototype: HTMLPosAddNewThingElement;
         new (): HTMLPosAddNewThingElement;
     };
+    interface HTMLPosAppElementEventMap {
+        "pod-os:session-changed": LoginEvent | LogoutEvent;
+        "pod-os:login": LoginEvent;
+        "pod-os:logout": LogoutEvent;
+    }
     interface HTMLPosAppElement extends Components.PosApp, HTMLStencilElement {
+        addEventListener<K extends keyof HTMLPosAppElementEventMap>(type: K, listener: (this: HTMLPosAppElement, ev: PosAppCustomEvent<HTMLPosAppElementEventMap[K]>) => any, options?: boolean | AddEventListenerOptions): void;
+        addEventListener<K extends keyof DocumentEventMap>(type: K, listener: (this: Document, ev: DocumentEventMap[K]) => any, options?: boolean | AddEventListenerOptions): void;
+        addEventListener<K extends keyof HTMLElementEventMap>(type: K, listener: (this: HTMLElement, ev: HTMLElementEventMap[K]) => any, options?: boolean | AddEventListenerOptions): void;
+        addEventListener(type: string, listener: EventListenerOrEventListenerObject, options?: boolean | AddEventListenerOptions): void;
+        removeEventListener<K extends keyof HTMLPosAppElementEventMap>(type: K, listener: (this: HTMLPosAppElement, ev: PosAppCustomEvent<HTMLPosAppElementEventMap[K]>) => any, options?: boolean | EventListenerOptions): void;
+        removeEventListener<K extends keyof DocumentEventMap>(type: K, listener: (this: Document, ev: DocumentEventMap[K]) => any, options?: boolean | EventListenerOptions): void;
+        removeEventListener<K extends keyof HTMLElementEventMap>(type: K, listener: (this: HTMLElement, ev: HTMLElementEventMap[K]) => any, options?: boolean | EventListenerOptions): void;
+        removeEventListener(type: string, listener: EventListenerOrEventListenerObject, options?: boolean | EventListenerOptions): void;
     }
     var HTMLPosAppElement: {
         prototype: HTMLPosAppElement;
@@ -721,6 +738,18 @@ declare namespace LocalJSX {
         "referenceUri": string;
     }
     interface PosApp {
+        /**
+          * Fires after a user has been authenticated successfully
+         */
+        "onPod-os:login"?: (event: PosAppCustomEvent<LoginEvent>) => void;
+        /**
+          * Fires after a user signed out
+         */
+        "onPod-os:logout"?: (event: PosAppCustomEvent<LogoutEvent>) => void;
+        /**
+          * Fired whenever the session login state changes (login / logout)
+         */
+        "onPod-os:session-changed"?: (event: PosAppCustomEvent<LoginEvent | LogoutEvent>) => void;
     }
     interface PosAppBrowser {
     }