+++ elements/src/components/pos-app/pos-app.spec.tsx
@@ -4,6 +4,7 @@ jest.mock('../../pod-os', () => ({
 
 import { newSpecPage } from '@stencil/core/testing';
 import { fireEvent } from '@testing-library/dom';
+import { when } from 'jest-when';
 import { PosApp } from './pos-app';
 import { createPodOS } from '../../pod-os';
 
@@ -22,34 +23,164 @@ describe('pos-app', () => {
       });
     });
 
-    it('does not load the preferences before login', async () => {
-      await newSpecPage({
-        components: [PosApp],
-        html: `<pos-app>item body</pos-app>`,
-        supportsShadowDom: false,
+    describe('load preferences', () => {
+      it('does not load the preferences before login', async () => {
+        await newSpecPage({
+          components: [PosApp],
+          html: `<pos-app>item body</pos-app>`,
+          supportsShadowDom: false,
+        });
+
+        trackSessionCallback({
+          isLoggedIn: false,
+          webId: 'https://pod.test/alice#me',
+        });
+
+        expect(mockFetchProfile).not.toHaveBeenCalled();
       });
 
-      trackSessionCallback({
-        isLoggedIn: false,
-        webId: 'https://pod.test/alice#me',
-      });
+      it('loads the preferences after login', async () => {
+        await newSpecPage({
+          components: [PosApp],
+          html: `<pos-app>item body</pos-app>`,
+          supportsShadowDom: false,
+        });
 
-      expect(mockFetchProfile).not.toHaveBeenCalled();
+        trackSessionCallback({
+          isLoggedIn: true,
+          webId: 'https://pod.test/alice#me',
+        });
+
+        expect(mockFetchProfile).toHaveBeenCalledWith('https://pod.test/alice#me');
+      });
     });
 
-    it('loads the preferences after login', async () => {
-      await newSpecPage({
-        components: [PosApp],
-        html: `<pos-app>item body</pos-app>`,
-        supportsShadowDom: false,
+    describe('communicates session changes', () => {
+      describe('on login', () => {
+        it('fires session-changed event with user data', async () => {
+          const onSessionChanged = jest.fn();
+          const page = await newSpecPage({
+            components: [PosApp],
+            html: `<pos-app>item body</pos-app>`,
+            supportsShadowDom: false,
+          });
+
+          when(mockFetchProfile).calledWith('https://pod.test/alice#me').mockResolvedValue('mocked WebIdProfile');
+
+          page.root.addEventListener('pod-os:session-changed', onSessionChanged);
+
+          trackSessionCallback(
+            {
+              isLoggedIn: true,
+              webId: 'https://pod.test/alice#me',
+            },
+            'mock authenticated fetch',
+          );
+
+          await page.waitForChanges();
+
+          expect(onSessionChanged).toHaveBeenCalledWith(
+            expect.objectContaining({
+              detail: {
+                isLoggedIn: true,
+                webId: 'https://pod.test/alice#me',
+                profile: 'mocked WebIdProfile',
+                authenticatedFetch: 'mock authenticated fetch',
+              },
+            }),
+          );
+        });
+
+        it('fires login event', async () => {
+          const onLogin = jest.fn();
+          const page = await newSpecPage({
+            components: [PosApp],
+            html: `<pos-app>item body</pos-app>`,
+            supportsShadowDom: false,
+          });
+
+          when(mockFetchProfile).calledWith('https://pod.test/alice#me').mockResolvedValue('mocked WebIdProfile');
+
+          page.root.addEventListener('pod-os:login', onLogin);
+
+          trackSessionCallback(
+            {
+              isLoggedIn: true,
+              webId: 'https://pod.test/alice#me',
+            },
+            'mock authenticated fetch',
+          );
+
+          await page.waitForChanges();
+
+          expect(onLogin).toHaveBeenCalledWith(
+            expect.objectContaining({
+              detail: {
+                isLoggedIn: true,
+                webId: 'https://pod.test/alice#me',
+                profile: 'mocked WebIdProfile',
+                authenticatedFetch: 'mock authenticated fetch',
+              },
+            }),
+          );
+        });
       });
 
-      trackSessionCallback({
-        isLoggedIn: true,
-        webId: 'https://pod.test/alice#me',
-      });
+      describe('on logout', () => {
+        it('fires session-changed event', async () => {
+          const onSessionChanged = jest.fn();
+          const page = await newSpecPage({
+            components: [PosApp],
+            html: `<pos-app>item body</pos-app>`,
+            supportsShadowDom: false,
+          });
 
-      expect(mockFetchProfile).toHaveBeenCalledWith('https://pod.test/alice#me');
+          when(mockFetchProfile).calledWith('https://pod.test/alice#me').mockResolvedValue('mocked WebIdProfile');
+
+          page.root.addEventListener('pod-os:session-changed', onSessionChanged);
+
+          trackSessionCallback({
+            isLoggedIn: false,
+          });
+
+          await page.waitForChanges();
+
+          expect(onSessionChanged).toHaveBeenCalledWith(
+            expect.objectContaining({
+              detail: {
+                isLoggedIn: false,
+              },
+            }),
+          );
+        });
+
+        it('fires logout event', async () => {
+          const onLogout = jest.fn();
+          const page = await newSpecPage({
+            components: [PosApp],
+            html: `<pos-app>item body</pos-app>`,
+            supportsShadowDom: false,
+          });
+
+          when(mockFetchProfile).calledWith('https://pod.test/alice#me').mockResolvedValue('mocked WebIdProfile');
+
+          page.root.addEventListener('pod-os:logout', onLogout);
+
+          trackSessionCallback({
+            isLoggedIn: false,
+          });
+
+          await page.waitForChanges();
+
+          expect(onLogout).toHaveBeenCalledWith(
+            expect.objectContaining({
+              detail: {
+                isLoggedIn: false,
+              },
+            }),
+          );
+        });
+      });
     });
 
     describe('load module', () => {