+++ elements/CHANGELOG.md
@@ -6,7 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
-- [pos-app](../docs/elements/components/pos-app): Can now handle new events of type `pod-os:module`, to asynchronously load a module and pass it to the component that triggered the event. Currently only the `contacts` module is supported.
+- [pos-app](../docs/elements/components/pos-app):
+  - Can now handle new events of type `pod-os:module`, to asynchronously load a module and pass it to the component that triggered the event. Currently only the `contacts` module is supported.
+  - Fires new events `pod-os:session-changed`, `pod-os:login` and `pod-os:logout` so that apps can react to changes in session state.
 
 ## 0.16.0